@@ -13,14 +13,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.0",
-    "@angular/common": "^17.3.0",
-    "@angular/compiler": "^17.3.0",
-    "@angular/core": "^17.3.0",
-    "@angular/forms": "^17.3.12",
-    "@angular/platform-browser": "^17.3.0",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/animations": "^18.2.3",
+    "@angular/common": "^18.2.3",
+    "@angular/compiler": "^18.2.3",
+    "@angular/core": "^18.2.3",
+    "@angular/forms": "^18.2.3",
+    "@angular/platform-browser": "^18.2.3",
+    "@angular/platform-browser-dynamic": "^18.2.3",
+    "@angular/router": "^18.2.3",
     "@auth0/angular-jwt": "^5.2.0",
     "angular-datatables": "^17.1.0",
     "bootstrap": "^5.3.3",
@@ -35,15 +35,15 @@
     "zone.js": "^0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.9",
+    "@angular-devkit/build-angular": "^18.2.3",
     "@angular-eslint/builder": "17.3.0",
     "@angular-eslint/eslint-plugin": "17.3.0",
     "@angular-eslint/eslint-plugin-template": "17.3.0",
     "@angular-eslint/schematics": "17.3.0",
     "@angular-eslint/template-parser": "17.3.0",
-    "@angular/cli": "~17.3.0",
-    "@angular/compiler-cli": "^17.3.4",
-    "@angular/localize": "^17.3.12",
+    "@angular/cli": "~18.2.3",
+    "@angular/compiler-cli": "^18.2.3",
+    "@angular/localize": "^18.2.3",
     "@types/bootstrap": "^5.2.10",
     "@types/datatables.net": "~1.12.0",
     "@types/jasmine": "~5.1.4",
