@@ -2,7 +2,7 @@ import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { EnsureModuleLoadedOnceGuard } from './guards/ensure-module-loaded-once.guard';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { AuthService } from './services/auth.service';
 import { AuthInterceptor } from './interceptors/auth.interceptor';
 import { TokenStorageService } from './services/token-storage.service';
@@ -10,20 +10,25 @@ import { ErrorInterceptor } from './interceptors/error.interceptor';
 import { JWT_OPTIONS, JwtHelperService } from '@auth0/angular-jwt';
 import { InvitesService } from './services/invites.service';
 
-@NgModule({
-  imports: [CommonModule, RouterModule, HttpClientModule],
-  exports: [RouterModule, HttpClientModule],
+@NgModule({ 
+  exports: [
+    RouterModule
+  ], 
+  imports: [
+    CommonModule, 
+    RouterModule], 
   providers: [
-    InvitesService,
-    AuthService,
-    TokenStorageService,
-    { provide: 'Window', useFactory: () => window },
-    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
-    { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
-    { provide: JWT_OPTIONS, useValue: JWT_OPTIONS },
-    JwtHelperService
-  ] // these should be singleton
-})
+      InvitesService,
+      AuthService,
+      TokenStorageService,
+      { provide: 'Window', useFactory: () => window },
+      { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
+      { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
+      { provide: JWT_OPTIONS, useValue: JWT_OPTIONS },
+      JwtHelperService,
+      provideHttpClient(withInterceptorsFromDi())
+    ] // these should be singleton
+ })
 export class CoreModule extends EnsureModuleLoadedOnceGuard {    // Ensure that CoreModule is only loaded into AppModule
   // Looks for the module in the parent injector to see if it's already been loaded (only want it loaded once)
   constructor(@Optional() @SkipSelf() parentModule: CoreModule) {
